tidy

+++ system/templates/deployment/global/home/start.bash
@@ -127,17 +127,20 @@ elif if test -f /etc/nginx
 	  then
 	 	mkdir /var/log/nginx
 	 fi
-  cp /home/ruby_env /home/.env_vars
-   for env_name in `cat /home/app.env `
-  	 do
-   		if ! test -z  "${!env_name}"
-         then
-        	 #val=`echo ${!env_name} | sed "/ /s//\\ /g"`
-  	      	 echo  "passenger_env_var $env_name \"${!env_name}\";"   >> /home/.env_vars
-  	     fi
-  	done
- echo " passenger_env_var RAILS_ENV $RAILS_ENV;" >> /home/.env_vars
- echo " passenger_env_var SECRET_KEY_BASE $SECRET_KEY_BASE;" >> /home/.env_vars
+     if test -f /home/ruby_env 
+      then
+       cp /home/ruby_env /home/.env_vars
+        for env_name in `cat /home/app.env `
+  	     do
+   		   if ! test -z  "${!env_name}"
+            then
+        	  #val=`echo ${!env_name} | sed "/ /s//\\ /g"`
+  	      	  echo  "passenger_env_var $env_name \"${!env_name}\";"   >> /home/.env_vars
+  	       fi
+  	     done
+       echo " passenger_env_var RAILS_ENV $RAILS_ENV;" >> /home/.env_vars
+       echo " passenger_env_var SECRET_KEY_BASE $SECRET_KEY_BASE;" >> /home/.env_vars
+	fi
 	
 	if test -f /home/engines/scripts/blocking.sh 
 	 then
@@ -149,11 +152,11 @@ elif if test -f /etc/nginx
 		nginx &
 	    echo  " $!" >>  $PID_FILE
 	fi
+	
 elif test -f /home/engines/scripts/blocking.sh 
-		then
-		   /home/engines/scripts/blocking.sh  &
-		   echo  " $!" >>  $PID_FILE
-		   
+	 then
+		/home/engines/scripts/blocking.sh  &
+	    echo  " $!" >>  $PID_FILE		   
 else
  echo "Nothing else to run!"
  exit
